Simplify UISettingsInjector
