+++ vault-ui/src/components/page/index.tsx
@@ -1,26 +1,23 @@
-import React, { PropsWithChildren, useEffect, useRef } from "react";
+import React, { PropsWithChildren } from "react";
 import { Outlet } from "react-router";
 
-import { useDarkMode } from "@/lib/dark-mode";
-import {
-  ProjectSettingsProvider,
-  useProjectSettings,
-} from "@/lib/project-settings";
+import { ProjectSettingsProvider } from "@/lib/project-settings";
 
+import { UISettingsInjector } from "../core/UISettingsInjector";
 import { SidebarInset, SidebarProvider } from "../ui/sidebar";
 import { VaultSidebar } from "./VaultSidebar";
 
 export function Page() {
   return (
     <ProjectSettingsProvider>
-      <PageOuter>
+      <UISettingsInjector>
         <SidebarProvider>
           <VaultSidebar />
           <SidebarInset>
             <PageInner />
           </SidebarInset>
         </SidebarProvider>
-      </PageOuter>
+      </UISettingsInjector>
     </ProjectSettingsProvider>
   );
 }
@@ -31,38 +28,6 @@ function PageInner() {
       <Outlet />
     </div>
   );
-}
-
-function PageOuter({ children }: PropsWithChildren) {
-  const rootRef = useRef<HTMLDivElement>(
-    document.getElementById("react-root") as HTMLDivElement,
-  );
-  const darkMode = useDarkMode();
-  const projectSettings = useProjectSettings();
-
-  useEffect(() => {
-    if (rootRef.current) {
-      if (darkMode) {
-        rootRef.current.classList.add("dark");
-        if (projectSettings?.darkModePrimaryColor) {
-          rootRef.current.style.setProperty(
-            "--primary",
-            projectSettings.darkModePrimaryColor,
-          );
-        }
-      } else {
-        rootRef.current.classList.remove("dark");
-        if (projectSettings?.primaryColor) {
-          rootRef.current.style.setProperty(
-            "--primary",
-            projectSettings.primaryColor,
-          );
-        }
-      }
-    }
-  }, [darkMode, projectSettings, rootRef]);
-
-  return <>{children}</>;
 }
 
 export function PageContent({ children }: PropsWithChildren) {
